shanoir-issue#2601: add new method to prevent duplicate of code

+++ shanoir-ng-datasets/src/main/java/org/shanoir/ng/datasetacquisition/service/DatasetAcquisitionServiceImpl.java
@@ -191,16 +191,18 @@ public class DatasetAcquisitionServiceImpl implements DatasetAcquisitionService 
         return updatedAcqs;
     }
 
-    @Override
-    @Transactional
-    public void deleteById(Long id, ShanoirEvent event) throws EntityNotFoundException, ShanoirException, SolrServerException, IOException, RestServiceException {
-        final DatasetAcquisition entity = repository.findById(id).orElse(null);
-        if (entity == null) {
-            throw new EntityNotFoundException("Cannot find entity with id = " + id);
-        }
-
+    /**
+     * Deletes children datasets then delete current acquisition
+     * @param entity
+     * @param event
+     * @throws ShanoirException
+     * @throws SolrServerException
+     * @throws IOException
+     * @throws RestServiceException
+     */
+    private void delete(DatasetAcquisition entity, ShanoirEvent event) throws ShanoirException, SolrServerException, IOException, RestServiceException {
         // Do not delete entity if it is the source. If getSourceId() is not null, it means it's a copy
-        List<DatasetAcquisition> childDsAc = repository.findBySourceId(id);
+        List<DatasetAcquisition> childDsAc = repository.findBySourceId(entity.getId());
         if (!CollectionUtils.isEmpty(childDsAc)) {
             throw new RestServiceException(
                     new ErrorModel(
@@ -224,50 +226,53 @@ public class DatasetAcquisitionServiceImpl implements DatasetAcquisitionService 
                 }
                 if (!datasetIds.isEmpty()) solrService.deleteFromIndex(datasetIds);
             }
-            String seriesInstanceUID = studyInstanceUIDHandler.findStudyInstanceUID(entity.getExamination());
-            String studyInstanceUID = seriesInstanceUIDHandler.findSeriesInstanceUID(entity);
-            dicomWebService.rejectAcquisitionFromPacs(seriesInstanceUID, studyInstanceUID);
-
-            repository.deleteById(id);
-            shanoirEventService.publishEvent(new ShanoirEvent(ShanoirEventType.DELETE_DATASET_ACQUISITION_EVENT, id.toString(), KeycloakUtil.getTokenUserId(), "", ShanoirEvent.SUCCESS, entity.getExamination().getStudyId()));
         }
     }
 
-    public void deleteByIdCascade(Long id, ShanoirEvent event) throws EntityNotFoundException, ShanoirException, SolrServerException, IOException, RestServiceException {
+    /**
+     * Call by acquisition-details > delete. Also reject current acquisition from pacs
+     * @param id
+     * @param event
+     * @throws ShanoirException
+     * @throws SolrServerException
+     * @throws IOException
+     * @throws RestServiceException
+     */
+    @Override
+    @Transactional
+    public void deleteById(Long id, ShanoirEvent event) throws ShanoirException, SolrServerException, IOException, RestServiceException {
         final DatasetAcquisition entity = repository.findById(id).orElse(null);
         if (entity == null) {
             throw new EntityNotFoundException("Cannot find entity with id = " + id);
         }
+        delete(entity, event);
 
-        // Do not delete entity if it is the source. If getSourceId() is not null, it means it's a copy
-        List<DatasetAcquisition> childDsAc = repository.findBySourceId(id);
-        if (!CollectionUtils.isEmpty(childDsAc)) {
-            throw new RestServiceException(
-                    new ErrorModel(
-                            HttpStatus.UNPROCESSABLE_ENTITY.value(),
-                            "This datasetAcquisition is linked to another datasetAcquisition that was copied."
-                    ));
-        } else {
-            List<Dataset> datasets = entity.getDatasets();
-            if (datasets != null) {
-                List<Long> datasetIds = new ArrayList<>();
-                for (Dataset ds : datasets) {
-                    if (event != null) {
-                        event.setMessage("Delete examination - dataset with id : " + ds.getId());
-                        float progressMax = Float.valueOf(event.getEventProperties().get("progressMax"));
-                        event.setProgress(event.getProgress() + (1f / progressMax));
-                        shanoirEventService.publishEvent(event);
-                    }
+        String seriesInstanceUID = studyInstanceUIDHandler.findStudyInstanceUID(entity.getExamination());
+        String studyInstanceUID = seriesInstanceUIDHandler.findSeriesInstanceUID(entity);
+        dicomWebService.rejectAcquisitionFromPacs(seriesInstanceUID, studyInstanceUID);
 
-                    datasetIds.add(ds.getId());
-                    datasetService.deleteByIdCascade(ds.getId());
-                }
-                if (!datasetIds.isEmpty()) solrService.deleteFromIndex(datasetIds);
-            }
-            repository.deleteById(id);
+        repository.deleteById(id);
+        shanoirEventService.publishEvent(new ShanoirEvent(ShanoirEventType.DELETE_DATASET_ACQUISITION_EVENT, id.toString(), KeycloakUtil.getTokenUserId(), "", ShanoirEvent.SUCCESS, entity.getExamination().getStudyId()));
+    }
 
-            shanoirEventService.publishEvent(new ShanoirEvent(ShanoirEventType.DELETE_DATASET_ACQUISITION_EVENT, id.toString(), KeycloakUtil.getTokenUserId(), "", ShanoirEvent.SUCCESS, entity.getExamination().getStudyId()));
+    /**
+     * Called by examination delete. Does not call reject from pacs as examination delete already does it
+     * @param id
+     * @param event
+     * @throws ShanoirException
+     * @throws SolrServerException
+     * @throws IOException
+     * @throws RestServiceException
+     */
+    public void deleteByIdCascade(Long id, ShanoirEvent event) throws ShanoirException, SolrServerException, IOException, RestServiceException {
+        final DatasetAcquisition entity = repository.findById(id).orElse(null);
+        if (entity == null) {
+            throw new EntityNotFoundException("Cannot find entity with id = " + id);
         }
+        delete(entity, event);
+
+        repository.deleteById(id);
+        shanoirEventService.publishEvent(new ShanoirEvent(ShanoirEventType.DELETE_DATASET_ACQUISITION_EVENT, id.toString(), KeycloakUtil.getTokenUserId(), "", ShanoirEvent.SUCCESS, entity.getExamination().getStudyId()));
     }
     
     @Override
